refactor: update to Tonga 0.16, add IPrologue Type

+++ src/WHyLL.Http/Warp/HttpWire.cs
@@ -126,7 +126,7 @@ namespace WHyLL.Http.Warp
                     new AsPrologue(
                         [
                             $"HTTP/{aspResponse.Version}",
-                            aspResponse.StatusCode.ToString(),
+                            ((int)aspResponse.StatusCode).ToString(),
                             aspResponse.ReasonPhrase
                         ]
                     ),
